refactor: twitter winner announcement feature
