+++ src/features/listings/components/Submission/index.tsx
@@ -1,1 +1,2 @@
+export * from './PublishResults';
 export * from './SubmissionActionButton';
