+++ src/features/listings/components/index.ts
@@ -2,5 +2,6 @@ export * from './ListingCard';
 export * from './ListingPage';
 export * from './ListingSection';
 export * from './ListingTabs';
+export * from './Submission';
 export * from './SubmissionsPage';
 export * from './WarningModal';
