+++ src/features/listings/utils/index.ts
@@ -99,3 +99,18 @@ export const getColorStyles = (status: string) => {
       return { bgColor: 'gray', color: 'white' };
   }
 };
+
+export function getBlobFromCanvas(
+  canvas: HTMLCanvasElement,
+  type: string,
+): Promise<Blob> {
+  return new Promise((resolve, reject) => {
+    canvas.toBlob((blob) => {
+      if (blob) {
+        resolve(blob);
+      } else {
+        reject(new Error('Blob creation failed'));
+      }
+    }, type);
+  });
+}
