--- src/utils/canvasToBlob.ts
@@ -1,14 +1,0 @@
-export function getBlobFromCanvas(
-  canvas: HTMLCanvasElement,
-  type: string,
-): Promise<Blob> {
-  return new Promise((resolve, reject) => {
-    canvas.toBlob((blob) => {
-      if (blob) {
-        resolve(blob);
-      } else {
-        reject(new Error('Blob creation failed'));
-      }
-    }, type);
-  });
-}
